Fix a sample time bug in the arc system model:
This bug is located in the Unit Delay block within the arc system model.
The purpose of this block is to prevent algebaric loops.

The previous `Sample time` parameter was `-1`, which made the sample
time inherit from the RMS block. So, the `Sample time` parameter was
equal to `1e-3`. This artificially injected a large delay in the arc current.
Further more, this delay resulted in the oscillation of the arc current
and the arc impedance RMSs.

In fact, a slight delay is enought to prevent algebaric loops. In this
case, the `Sample time` parameter is set to `1e-5`, which is the same as
the simulation step size.